requirements_gpu.txt
scikit-learn==0.21.3
keras==2.3.1
Pillow
tensorflow-gpu==1.14.0
h5py==2.10.0
imutils==0.5.3
opencv-contrib-python==4.1.0.25
progressbar==2.5
matplotlib